move implementation from types module
